+++ test/system/super_scaffolding/partial_test/teardown.rb
@@ -1,0 +1,17 @@
+#!/usr/bin/env ruby
+
+require_relative "../super_scaffolding_test_teardown"
+
+class Teardown < SuperScaffoldingTestTeardown
+  # This allows us to define helper methods that aren't attached to thor commands
+  no_commands do
+    def teardown
+      puts "do the thing"
+    end
+  end
+end
+
+# We create our own args array so that we don't have to ask the user to include `scaffolding_setup` on the command line
+args = ["scaffolding_teardown"] + ARGV
+
+Teardown.start(args)
